Bug fixes and changes
This change fixes bugs related to broken mobile search interface by removing backdrop filter on mobile size by working around and broken mobile navigation when viewing blog posts.
This also changes the links that was pointing to from blog to blogs, change the behavior from reloading to force reload when clicked the link that meant to visit blog on the footer, adjust Z index on that and make paragraph weight bolder.